% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IoAFUN.R
\name{IoA}
\alias{IoA}
\title{Index of Agreement over a running window function}
\usage{
IoA(x, obs, comparison.value, date, window.size)
}
\arguments{
\item{obs}{The column under investigation.}

\item{comparison.value}{The comparison column.}

\item{date}{Date column.}

\item{window.size}{The size of the running window.}
}
\description{
Index of Agreement over a running window function
}
\examples{
IoAFUN()
}
