% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualInfoFUN.R
\name{mutualInfoFUN}
\alias{mutualInfoFUN}
\title{Mutual Information function.}
\usage{
mutualInfoFUN(x, obs, comparison.value, bin = 5)
}
\arguments{
\item{obs}{The first column under investigation.}

\item{comparison.value}{The second column.}

\item{bin}{Bin size for grouping the data. Default is 5.}
}
\description{
Checks the mutual information between two samples using the Kullback-Leibler divergence in the `entropy` package.
}
\examples{
mutualInfoFUN()
}
