% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runningSeqFUN.R
\name{runningSeqFUN}
\alias{runningSeqFUN}
\title{Running Sequence Function for Blocks of Data Analysis.}
\usage{
runningSeqFUN(x, date, run.length, measurement.freq)
}
\arguments{
\item{date}{Date column.}

\item{run.length}{The size of a block of data to analyse.}

\item{measurement.freq}{Either `mins` (default) or `hours`.}
}
\description{
A function for labelling data into their blocks.
}
\examples{
runningSeqFUN()
}
