% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmailFUN.R
\name{sendEmailFUN}
\alias{sendEmailFUN}
\title{send email in R.}
\usage{
sendEmailFUN(email.title, from, to, body, from.passwd,
  host.name = "smtp.gmail.com", host.port = 465)
}
\arguments{
\item{email.title}{Email subject.}

\item{from}{Email address being sent from. This will need to be configured to allow for emails to be sent from third parties.}

\item{to}{Email adress to send to.}

\item{body}{Body of email. Can be images.}

\item{from.passwd}{The password to the `from` email address.}

\item{host.name}{The host of the `from` email. The default here is for  Google email address.}

\item{host.port}{The host port. Default is 465.}
}
\description{
This function uses the `mailR` package to send an email from R. I normally use this function at the end of some analysis, where the body is the end result.
}
\examples{
sendEmailFUN()
}
